issue #155 Removed get/setDynamicData methods from AbstractEntity added these methods to Question removed null methods from Exam and modified ExamTest
