+++ src/main/java/com/haxwell/apps/questions/entities/AbstractEntity.java
@@ -19,57 +19,13 @@ package com.haxwell.apps.questions.entities;
  * along with Quizki. If not, see http://www.gnu.org/licenses.
  */
 
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Set;
 
-import net.minidev.json.JSONObject;
 
 public abstract class AbstractEntity {
-
-	/*
-	 * The interface Comparable was never implemented (Eclipse can't resolve it) and there are no other references to it. Questions are
-	 * compared but no other entities are, or would make sense to be. Choices might if we want to sort them into an order but it's still
-	 * not a generic requirement for all entity types so I removed this.
-	 */
-	
-	
 	
 	protected final boolean APPEND_COMMA = true;
 	
-	protected Map<String, Object> dynamicData = new HashMap<String, Object>();
 
-	/*	
-	 * removed @Override
-	 * This might have been related to the planned (or removed) Comparable interface that was never implemented.
-	 * 
-	 * Jonathan I'm planning on removing compareTo() from here. It's only used to compare Questions in 6 places in the project. None of
-	 * the other entities use it. Since Question does implement it simply removing it here will not affect it or any of the other entities. Is
-	 * that correct (besides the Comparable interface problem)? It is commented out here to test that, but I am not sure the test suite would catch it.
-	 * 
-	 * How can method removals be tested?
-	 * 
-	 * I resolved the problem with the @Override related compilation error in this version.
-	 * 
-	 * (In a separate issue I am getting an error from maven "Could not find a value for QUIZKI_JDBC_URL in System.getProperty()" that is preventing
-	 * the test suite from connecting to the database when I build the project. Where is that variable set and to what?)
-	 * 
-	 * Thank you in advance for any help on this
-	 
-	
-	public int compareTo(Object o) {
-		return 0;
-	}
-*/
-	public Object getDynamicData(String key) {
-		return dynamicData.get(key);
-	}
-	
-	public void setDynamicData(String key, Object o) {
-		dynamicData.put(key, o);
-	}
-	
 	public long getId() {
 		return Long.MIN_VALUE;
 	}
@@ -90,13 +46,5 @@ public abstract class AbstractEntity {
 		return null;
 	}
 
-	protected void addDynamicDataToJSONObject(JSONObject j) {
-		Set<String> keys = dynamicData.keySet();
-		
-		j.put("dynamicDataFieldNames", keys.toArray());
-		
-		for(String key : keys) {
-			j.put(key,  dynamicData.get(key).toString());
-		}
-	}
+	
 }
